strcmp

+++ src/libft/ft_strcmp.c
@@ -6,13 +6,15 @@
 /*   By: nalexand <nalexand@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/04/02 18:16:18 by nalexand          #+#    #+#             */
-/*   Updated: 2019/05/23 07:45:34 by nalexand         ###   ########.fr       */
+/*   Updated: 2019/07/28 03:50:58 by nalexand         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
+#include "libft.h"
+
 int		ft_strcmp(const char *s1, const char *s2)
 {
-	int i;
+	size_t	 i;
 
 	i = 0;
 	while (s1[i] || s2[i])
